Move `StringBuilder` to its own module
